Update simpleGrid.js
Adding to the search criteria for the dropdowns, as it's failing in some instances - seems to be going with a regular search, wherein it'll pull all words rather than an exact phrase.  E.g. "C78 abc 7" and "C78 abc 8" will always return both, as they match 7 and 8, C, and abc.